Migrated tooltip action to svelte 5
