+++ src/actions/index.ts
@@ -1,1 +1,1 @@
-export { default as tooltip } from './tooltip';
+export { default as tooltip } from './tooltip.svelte';
